refactor: split detection logic
