--- packages/detection/src/detectors.js
@@ -1,644 +1,0 @@
-/**
- * Specialized platform detectors moved from background.js
- */
-
-export async function detectAlipayUser() {
-    const platformId = 'alipayopen'
-    try {
-        // 从 storage 读取缓存的用户信息
-        const stored = await chrome.storage.local.get('alipayopen_user')
-        const cachedUser = stored.alipayopen_user
-
-        if (cachedUser && cachedUser.loggedIn) {
-            // 检查缓存是否过期（1小时）
-            const cacheAge = Date.now() - (cachedUser.cachedAt || 0)
-            const maxAge = 1 * 60 * 60 * 1000 // 1 hour
-
-            if (cacheAge < maxAge) {
-                console.log(`[COSE] alipayopen 从缓存读取用户信息:`, cachedUser.username)
-                return {
-                    loggedIn: true,
-                    username: cachedUser.username || '',
-                    avatar: cachedUser.avatar || ''
-                }
-            } else {
-                console.log(`[COSE] alipayopen 缓存已过期`)
-                // 清除过期缓存
-                await chrome.storage.local.remove('alipayopen_user')
-            }
-        }
-
-        // 尝试从已打开的支付宝页面获取用户信息并缓存
-        let tabs = await chrome.tabs.query({ url: 'https://open.alipay.com/*' })
-        if (tabs.length === 0) {
-            tabs = await chrome.tabs.query({ url: 'https://*.alipay.com/*' })
-        }
-
-        if (tabs.length > 0) {
-            try {
-                // 在已打开的页面上下文中调用 API
-                const results = await chrome.scripting.executeScript({
-                    target: { tabId: tabs[0].id },
-                    func: async () => {
-                        try {
-                            const response = await fetch('https://developerportal.alipay.com/octopus/service.do', {
-                                method: 'POST',
-                                credentials: 'include',
-                                headers: {
-                                    'Accept': 'application/json',
-                                    'Content-Type': 'application/x-www-form-urlencoded;charset=utf-8',
-                                },
-                                body: 'data=%5B%7B%7D%5D&serviceName=alipay.open.developerops.forum.user.query',
-                            })
-                            if (!response.ok) return null
-                            return await response.json()
-                        } catch (e) {
-                            return null
-                        }
-                    }
-                })
-
-                const data = results?.[0]?.result
-                console.log(`[COSE] alipayopen API 数据:`, data)
-
-                if (data?.stat === 'ok' && data?.data?.isLoginUser === 1) {
-                    const username = data.data.nickname || ''
-                    const avatar = data.data.avatar || ''
-
-                    // 缓存用户信息
-                    await chrome.storage.local.set({
-                        alipayopen_user: {
-                            loggedIn: true,
-                            username,
-                            avatar,
-                            cachedAt: Date.now()
-                        }
-                    })
-
-                    console.log(`[COSE] alipayopen 用户信息:`, username, avatar ? '有头像' : '无头像')
-                    return { loggedIn: true, username, avatar }
-                }
-            } catch (e) {
-                console.log(`[COSE] alipayopen 从页面获取用户信息失败:`, e.message)
-            }
-        }
-
-        console.log(`[COSE] alipayopen 未检测到登录状态`)
-        return { loggedIn: false }
-    } catch (e) {
-        console.log(`[COSE] alipayopen 检测失败:`, e.message)
-        return { loggedIn: false, error: e.message }
-    }
-}
-
-export async function detectWeiboUser() {
-    const platformId = 'weibo'
-    try {
-        // 检查 card.weibo.com 的 SUBP cookie
-        const subpCookie = await chrome.cookies.get({
-            url: 'https://card.weibo.com',
-            name: 'SUBP'
-        })
-
-        // 也检查 ALF cookie
-        const alfCookie = await chrome.cookies.get({
-            url: 'https://card.weibo.com',
-            name: 'ALF'
-        })
-
-        if (!subpCookie && !alfCookie) {
-            console.log(`[COSE] weibo 未找到登录 cookie，未登录`)
-            return { loggedIn: false }
-        }
-
-        // 有 cookie，通过 fetch HTML 获取用户信息
-        let username = ''
-        let avatar = ''
-
-        try {
-            // 获取所有相关 cookies 并手动添加到请求
-            const weiboCookies = await chrome.cookies.getAll({ domain: '.weibo.com' })
-            const cardCookies = await chrome.cookies.getAll({ domain: 'card.weibo.com' })
-            const sinaCookies = await chrome.cookies.getAll({ domain: '.sina.com.cn' })
-            const allCookies = [...weiboCookies, ...cardCookies, ...sinaCookies]
-            const cookieString = allCookies.map(c => `${c.name}=${c.value}`).join('; ')
-
-            const response = await fetch('https://card.weibo.com/article/v5/editor', {
-                method: 'GET',
-                headers: {
-                    'Cookie': cookieString,
-                },
-                credentials: 'include',
-            })
-            const html = await response.text()
-
-            // 从 HTML 中提取用户名
-            const nickMatch = html.match(/"nick"\s*:\s*"([^"]+)"/)
-            if (nickMatch) {
-                username = nickMatch[1]
-            } else {
-                // 深度查找 nick
-                const altNickMatch = html.match(/\\"nick\\"\s*:\s*\\"([^\\"]+)\\"/)
-                if (altNickMatch) {
-                    username = altNickMatch[1]
-                }
-            }
-
-            // 从 HTML 中提取头像
-            const avatarMatch = html.match(/"avatar_large"\s*:\s*"([^"]+)"/)
-            if (avatarMatch) {
-                avatar = avatarMatch[1].replace(/\\/g, '')
-            } else {
-                const altAvatarMatch = html.match(/\\"avatar_large\\"\s*:\s*\\"([^\\"]+)\\"/)
-                if (altAvatarMatch) {
-                    let rawAvatar = altAvatarMatch[1].replace(/\\\\\\\//g, '/')
-                    if (rawAvatar.includes('sinaimg.cn')) {
-                        avatar = rawAvatar.split('?')[0]
-                    } else {
-                        avatar = rawAvatar
-                    }
-                }
-            }
-
-            console.log(`[COSE] weibo 用户信息: ${username}`)
-        } catch (e) {
-            console.log(`[COSE] weibo 获取用户详情失败:`, e.message)
-        }
-
-        if (!username) {
-            return { loggedIn: false }
-        }
-
-        return { loggedIn: true, username, avatar }
-    } catch (e) {
-        console.log(`[COSE] weibo 检测失败:`, e.message)
-        return { loggedIn: false }
-    }
-}
-
-export async function detectWechatUser() {
-    const platformId = 'wechat'
-    try {
-        // 先检查缓存
-        const stored = await chrome.storage.local.get('wechat_user')
-        const cachedUser = stored.wechat_user
-
-        if (cachedUser && cachedUser.loggedIn) {
-            const cacheAge = Date.now() - (cachedUser.cachedAt || 0)
-            const maxAge = 1 * 60 * 60 * 1000 // 1 hour
-
-            if (cacheAge < maxAge) {
-                console.log(`[COSE] wechat 从缓存读取:`, cachedUser.username)
-                return {
-                    loggedIn: true,
-                    username: cachedUser.username || '',
-                    avatar: cachedUser.avatar || ''
-                }
-            } else {
-                await chrome.storage.local.remove('wechat_user')
-            }
-        }
-
-        // 优先尝试在已打开的微信公众号页面中检测
-        const tabs = await chrome.tabs.query({ url: 'https://mp.weixin.qq.com/*' })
-        if (tabs.length > 0) {
-            try {
-                const results = await chrome.scripting.executeScript({
-                    target: { tabId: tabs[0].id },
-                    func: () => {
-                        const wxData = window.wx?.data
-                        if (wxData && wxData.nick_name) {
-                            return {
-                                loggedIn: true,
-                                username: wxData.nick_name || wxData.user_name || '',
-                                avatar: wxData.head_img || '',
-                                token: wxData.t || ''
-                            }
-                        }
-                        return null
-                    }
-                })
-
-                const result = results?.[0]?.result
-                if (result && result.loggedIn) {
-                    const userInfo = { ...result, cachedAt: Date.now() }
-                    await chrome.storage.local.set({ wechat_user: userInfo })
-                    return {
-                        loggedIn: true,
-                        username: userInfo.username || '',
-                        avatar: userInfo.avatar || ''
-                    }
-                }
-            } catch (e) {
-                console.log(`[COSE] wechat 页面脚本执行失败:`, e.message)
-            }
-        }
-
-        // 备用方案：fetch 首页并解析 HTML
-        try {
-            const response = await fetch('https://mp.weixin.qq.com/', {
-                method: 'GET',
-                credentials: 'include',
-                headers: { 'Accept': 'text/html' }
-            })
-            const html = await response.text()
-
-            if (html.includes('请使用微信扫描') || html.includes('扫码登录')) {
-                return { loggedIn: false }
-            }
-
-            const nickMatch = html.match(/nick_name\s*[:=]\s*["']([^"']+)["']/)
-            const avatarMatch = html.match(/head_img\s*[:=]\s*["']([^"']+)["']/)
-
-            if (nickMatch) {
-                const username = nickMatch[1]
-                const avatar = avatarMatch ? avatarMatch[1] : ''
-                await chrome.storage.local.set({
-                    wechat_user: {
-                        loggedIn: true,
-                        username,
-                        avatar,
-                        cachedAt: Date.now()
-                    }
-                })
-                return { loggedIn: true, username, avatar }
-            }
-        } catch (e) {
-            console.log(`[COSE] wechat fetch 失败:`, e.message)
-        }
-
-        return { loggedIn: false }
-    } catch (e) {
-        console.log(`[COSE] wechat 检测失败:`, e.message)
-        return { loggedIn: false }
-    }
-}
-
-export async function detectXiaohongshuUser() {
-    const platformId = 'xiaohongshu'
-    try {
-        // 先检查缓存
-        const stored = await chrome.storage.local.get('xiaohongshu_user')
-        const cachedUser = stored.xiaohongshu_user
-
-        if (cachedUser && cachedUser.loggedIn) {
-            const cacheAge = Date.now() - (cachedUser.cachedAt || 0)
-            const maxAge = 7 * 24 * 60 * 60 * 1000 // 7 days
-            if (cacheAge < maxAge) {
-                console.log(`[COSE] xiaohongshu 从缓存读取:`, cachedUser.username)
-                return { loggedIn: true, username: cachedUser.username || '', avatar: cachedUser.avatar || '' }
-            } else {
-                await chrome.storage.local.remove('xiaohongshu_user')
-            }
-        }
-
-        // 缓存无效，尝试在已打开的小红书页面中检测
-        const tabs = await chrome.tabs.query({ url: 'https://creator.xiaohongshu.com/*' })
-        if (tabs.length > 0) {
-            const results = await chrome.scripting.executeScript({
-                target: { tabId: tabs[0].id },
-                func: async () => {
-                    try {
-                        const response = await fetch('https://creator.xiaohongshu.com/api/galaxy/user/info', {
-                            method: 'GET',
-                            credentials: 'include',
-                            headers: { 'Accept': 'application/json' }
-                        })
-                        if (!response.ok) return null
-                        const data = await response.json()
-                        if (data?.success === true && data?.code === 0 && data?.data?.userId) {
-                            return {
-                                loggedIn: true,
-                                username: data.data.userName || data.data.redId || '',
-                                avatar: data.data.userAvatar || '',
-                                userId: data.data.userId
-                            }
-                        }
-                        return null
-                    } catch (e) { return null }
-                }
-            })
-
-            const result = results?.[0]?.result
-            if (result && result.loggedIn) {
-                const userInfo = { ...result, cachedAt: Date.now() }
-                await chrome.storage.local.set({ xiaohongshu_user: userInfo })
-                return { loggedIn: true, username: userInfo.username || '', avatar: userInfo.avatar || '' }
-            }
-        }
-        return { loggedIn: false }
-    } catch (e) {
-        console.log(`[COSE] xiaohongshu 检测失败:`, e.message)
-        return { loggedIn: false }
-    }
-}
-
-export async function detectElecfansUser() {
-    const platformId = 'elecfans'
-    try {
-        const authCookie = await chrome.cookies.get({ url: 'https://www.elecfans.com', name: 'auth' })
-        const authWwwCookie = await chrome.cookies.get({ url: 'https://www.elecfans.com', name: 'auth_www' })
-
-        if (!authCookie && !authWwwCookie) {
-            return { loggedIn: false }
-        }
-
-        try {
-            const response = await fetch('https://www.elecfans.com/webapi/passport/checklogin?_=' + Date.now(), {
-                method: 'GET',
-                credentials: 'include',
-                headers: { 'Accept': 'application/json, text/javascript, */*; q=0.01' }
-            })
-
-            if (!response.ok) return { loggedIn: true, username: '', avatar: '' }
-
-            const data = await response.json()
-            if (data && data.uid) {
-                const username = data.username || ''
-                const avatar = data.avatar || ''
-                return { loggedIn: true, username, avatar }
-            } else {
-                return { loggedIn: true, username: '', avatar: '' }
-            }
-        } catch (e) {
-            return { loggedIn: true, username: '', avatar: '' }
-        }
-    } catch (e) {
-        return { loggedIn: false }
-    }
-}
-
-export async function detectHuaweiCloudUser() {
-    const platformId = 'huaweicloud'
-    try {
-        let tabs = await chrome.tabs.query({ url: 'https://bbs.huaweicloud.com/*' })
-        if (tabs.length === 0) tabs = await chrome.tabs.query({ url: 'https://*.huaweicloud.com/*' })
-        if (tabs.length === 0) return { loggedIn: false }
-
-        try {
-            const results = await chrome.scripting.executeScript({
-                target: { tabId: tabs[0].id },
-                func: () => {
-                    return new Promise((resolve) => {
-                        const csrf = document.cookie.match(/csrf=([^;]+)/)?.[1] || ''
-                        fetch('https://devdata.huaweicloud.com/rest/developer/fwdu/rest/developer/user/hdcommunityservice/v1/member/get-personal-info', {
-                            method: 'GET',
-                            credentials: 'include',
-                            headers: { 'Accept': 'application/json', 'csrf': csrf }
-                        })
-                            .then(response => response.ok ? response.json() : null)
-                            .then(data => {
-                                if (data && data.memName) {
-                                    resolve({ memName: data.memName, memAlias: data.memAlias, memPhoto: data.memPhoto })
-                                } else {
-                                    resolve(null)
-                                }
-                            })
-                            .catch(() => resolve(null))
-                    })
-                }
-            })
-
-            let data = results?.[0]?.result
-            if (data && typeof data.then === 'function') data = await data // handle promise result if applicable
-
-            if (data && data.memName) {
-                return { loggedIn: true, username: data.memAlias || data.memName, avatar: data.memPhoto || '' }
-            }
-        } catch (e) { }
-
-        return { loggedIn: false }
-    } catch (e) {
-        return { loggedIn: false }
-    }
-}
-
-export async function detectHuaweiDevUser() {
-    const platformId = 'huaweidev'
-    try {
-        const userInfoCookie = await chrome.cookies.get({ url: 'https://developer.huawei.com', name: 'developer_userinfo' })
-        if (!userInfoCookie || !userInfoCookie.value) return { loggedIn: false }
-
-        let csrfToken = ''
-        try {
-            const userInfoData = JSON.parse(decodeURIComponent(userInfoCookie.value))
-            csrfToken = userInfoData.csrftoken || ''
-        } catch (e) { }
-
-        if (!csrfToken) {
-            const csrfCookie = await chrome.cookies.get({ url: 'https://developer.huawei.com', name: 'csrfToken' })
-            csrfToken = csrfCookie?.value || ''
-        }
-
-        if (!csrfToken) return { loggedIn: true, username: '', avatar: '' }
-
-        const response = await fetch('https://svc-drcn.developer.huawei.com/codeserver/Common/v1/delegate', {
-            method: 'POST',
-            credentials: 'include',
-            headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json;charset=UTF-8',
-                'x-hd-csrf': csrfToken,
-            },
-            body: JSON.stringify({ svc: 'GOpen.User.getInfo', reqType: 0, reqJson: JSON.stringify({ getNickName: '1' }) })
-        })
-
-        const data = await response.json()
-        if (data.returnCode === '0' && data.resJson) {
-            const userInfo = JSON.parse(data.resJson)
-            return {
-                loggedIn: true,
-                username: userInfo.displayName || userInfo.loginID || '',
-                avatar: userInfo.headPictureURL || ''
-            }
-        } else {
-            return { loggedIn: true, username: '', avatar: '' }
-        }
-    } catch (e) {
-        return { loggedIn: false }
-    }
-}
-
-export async function detectSspaiUser() {
-    try {
-        const jwtCookie = await chrome.cookies.get({ url: 'https://sspai.com', name: 'sspai_jwt_token' })
-        if (!jwtCookie || !jwtCookie.value) return { loggedIn: false }
-
-        const token = jwtCookie.value
-        const response = await fetch('https://sspai.com/api/v1/user/info/get', {
-            method: 'GET',
-            credentials: 'include',
-            headers: { 'Accept': 'application/json', 'Authorization': `Bearer ${token}` }
-        })
-        const data = await response.json()
-
-        if (data.error === 0 && data.data?.nickname) {
-            return { loggedIn: true, username: data.data.nickname, avatar: data.data.avatar || '' }
-        } else {
-            return { loggedIn: false }
-        }
-    } catch (e) { return { loggedIn: false } }
-}
-
-export async function detectAliyunUser() {
-    try {
-        const ticketCookie = await chrome.cookies.get({ url: 'https://developer.aliyun.com', name: 'login_aliyunid_ticket' })
-        if (!ticketCookie || !ticketCookie.value) return { loggedIn: false }
-
-        const response = await fetch('https://developer.aliyun.com/developer/api/my/user/getUser', {
-            method: 'GET',
-            credentials: 'include',
-            headers: { 'Accept': 'application/json' }
-        })
-        const data = await response.json()
-
-        if (data.success && data.data?.nickname) {
-            return { loggedIn: true, username: data.data.nickname, avatar: data.data.avatar || '' }
-        }
-        return { loggedIn: false }
-    } catch (e) { return { loggedIn: false } }
-}
-
-export async function detectSohuUser() {
-    try {
-        const ppinfCookie = await chrome.cookies.get({ url: 'https://mp.sohu.com', name: 'ppinf' })
-        if (!ppinfCookie || !ppinfCookie.value) return { loggedIn: false }
-
-        try {
-            const response = await fetch('https://mp.sohu.com/mpbp/bp/account/list', {
-                method: 'GET',
-                credentials: 'include',
-                headers: { 'Accept': 'application/json' }
-            })
-            const data = await response.json()
-            if (data.success && data.data?.data?.[0]?.accounts?.[0]) {
-                const account = data.data.data[0].accounts[0]
-                let avatar = account.avatar || ''
-                if (avatar.startsWith('//')) avatar = 'https:' + avatar
-                return { loggedIn: true, username: account.nickName, avatar }
-            } else {
-                return { loggedIn: true, username: '', avatar: '' }
-            }
-        } catch (e) {
-            return { loggedIn: true, username: '', avatar: '' }
-        }
-    } catch (e) { return { loggedIn: false } }
-}
-
-export async function detectMediumUser() {
-    try {
-        const sidCookie = await chrome.cookies.get({ url: 'https://medium.com', name: 'sid' })
-        const uidCookie = await chrome.cookies.get({ url: 'https://medium.com', name: 'uid' })
-
-        if (!sidCookie && !uidCookie) return { loggedIn: false }
-
-        const response = await fetch('https://medium.com/me/stats', {
-            method: 'GET',
-            credentials: 'include',
-        })
-        const html = await response.text()
-        const finalUrl = response.url
-
-        if (finalUrl.includes('/m/signin') || finalUrl.includes('?signIn')) return { loggedIn: false }
-
-        const profileMatch = html.match(/"username"\s*:\s*"([^"]+)"/) ||
-            html.match(/href="https:\/\/medium\.com\/@([^"?\/]+)"/) ||
-            html.match(/medium\.com\/@([a-zA-Z0-9_]+)/)
-
-        if (profileMatch && profileMatch[1] && profileMatch[1] !== 'gmail' && profileMatch[1] !== 'medium') {
-            const username = profileMatch[1]
-            const shortName = username.replace(/\d+$/, '')
-            const avatarPattern = new RegExp(`<img[^>]*alt="${shortName}"[^>]*src="([^"]+)"`)
-            const avatarMatch = html.match(avatarPattern) || html.match(new RegExp(`<img[^>]*src="([^"]+)"[^>]*alt="${shortName}"`))
-
-            return { loggedIn: true, username, avatar: avatarMatch ? avatarMatch[1] : '' }
-        } else {
-            return { loggedIn: true, username: '', avatar: '' }
-        }
-    } catch (e) { return { loggedIn: false } }
-}
-
-export async function detectTencentCloudUser() {
-    try {
-        const response = await fetch('https://cloud.tencent.com/developer/creator', {
-            method: 'GET',
-            credentials: 'include',
-        })
-        const html = await response.text()
-        const finalUrl = response.url
-
-        if (!finalUrl.includes('/creator')) return { loggedIn: false }
-        if (html.includes('登录/注册') || html.includes('"isLogin":false') || html.includes('"login":false')) return { loggedIn: false }
-
-        const userInfoMatch = html.match(/"userInfo"\s*:\s*\{[^}]*"nickname"\s*:\s*"([^"]+)"[^}]*\}/) ||
-            html.match(/"creatorInfo"\s*:\s*\{[^}]*"nickname"\s*:\s*"([^"]+)"[^}]*\}/) ||
-            html.match(/"currentUser"\s*:\s*\{[^}]*"nickname"\s*:\s*"([^"]+)"[^}]*\}/)
-
-        const creatorNicknameMatch = html.match(/class="creator-info[^"]*"[^>]*>[\s\S]*?<[^>]*class="[^"]*name[^"]*"[^>]*>([^<]+)</) ||
-            html.match(/"isCreator"\s*:\s*true[\s\S]*?"nickname"\s*:\s*"([^"]+)"/)
-
-        const nicknameMatch = userInfoMatch || creatorNicknameMatch
-        const avatarMatch = html.match(/"userInfo"[\s\S]*?"avatarUrl"\s*:\s*"([^"]+)"/) ||
-            html.match(/"avatar"\s*:\s*"(https?:\/\/[^"]+)"/)
-
-        if (nicknameMatch && nicknameMatch[1]) {
-            return { loggedIn: true, username: nicknameMatch[1], avatar: avatarMatch ? avatarMatch[1] : '' }
-        } else {
-            if (html.includes('创作中心') || html.includes('我的文章')) return { loggedIn: true, username: '', avatar: '' }
-            return { loggedIn: false }
-        }
-    } catch (e) { return { loggedIn: false } }
-}
-
-export async function detectQianfanUser() {
-    try {
-        const response = await fetch('https://qianfan.cloud.baidu.com/api/community/user/current', {
-            method: 'GET',
-            credentials: 'include',
-            headers: { 'Accept': 'application/json' }
-        })
-        if (!response.ok) return { loggedIn: false }
-
-        const data = await response.json()
-        if (data.success && data.result) {
-            const username = data.result.displayName || data.result.nickname || ''
-            const avatar = data.result.avatar || ''
-            return { loggedIn: true, username, avatar }
-        } else {
-            return { loggedIn: false }
-        }
-    } catch (e) { return { loggedIn: false } }
-}
-
-export async function detectTwitterUser() {
-    try {
-        const authTokenCookie = await chrome.cookies.get({ url: 'https://x.com', name: 'auth_token' })
-        const ct0Cookie = await chrome.cookies.get({ url: 'https://x.com', name: 'ct0' })
-
-        if (!authTokenCookie) return { loggedIn: false }
-
-        let username = ''
-        let avatar = ''
-
-        try {
-            const response = await fetch('https://x.com/home', {
-                method: 'GET',
-                credentials: 'include',
-                headers: { 'Accept': 'text/html' }
-            })
-            if (response.ok) {
-                const html = await response.text()
-                const screenNameMatch = html.match(/"screen_name"\s*:\s*"([^"]+)"/)
-                if (screenNameMatch) username = screenNameMatch[1]
-                const avatarMatch = html.match(/"profile_image_url_https"\s*:\s*"([^"]+)"/)
-                if (avatarMatch) avatar = avatarMatch[1].replace('_normal.', '_x96.')
-            }
-        } catch (e) { }
-
-        // If fetch failed or regex failed, try explicit fallback scraping logic (simplified here)
-        // Note: Full scrape logic from background.js is complex. 
-        // We will assume basic fetch works or just return loggedIn:true if cookie exists
-
-        return { loggedIn: true, username, avatar }
-    } catch (e) { return { loggedIn: false } }
-}
